pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "social-post"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = ["python-dotenv", "requests", "python-dateutil", "openai"]

[tool.setuptools.packages.find]
where = ["src"]
